feat: add warnings

+++ src/main.cpp
@@ -36,11 +36,12 @@ static void render_loop(ImGuiIO &io, SDL_Renderer *renderer, SDL_Window *window)
 
         // Rendering
         ImGui::Render();
+
         SDL_RenderSetScale(renderer, io.DisplayFramebufferScale.x,
                            io.DisplayFramebufferScale.y);
-        SDL_SetRenderDrawColor(
-            renderer, (Uint8)(clear_color.x * 255), (Uint8)(clear_color.y * 255),
-            (Uint8)(clear_color.z * 255), (Uint8)(clear_color.w * 255));
+
+        SDL_SetRenderDrawColor(renderer, clear_color.x * 255, clear_color.y * 255, clear_color.z * 255, clear_color.w * 255);
+
         SDL_RenderClear(renderer);
         ImGui_ImplSDLRenderer_RenderDrawData(ImGui::GetDrawData());
         SDL_RenderPresent(renderer);
